requirements.txt
crispy-forms-foundation==0.6.4
Django==1.10.8
django-autocomplete-light==3.2.10
django-bootstrap3==9.0.0
django-braces==1.11.0
django-classy-tags==0.8.0
django-crispy-forms==1.6.1
django-eztables==0.3.2
django-filter==1.0.4
django-sekizai==0.10.0
django-tables2==1.10.0
django.js==0.8.1
foundation==0.1.0a0.dev1
pytz==2017.2
six==1.10.0
